+++ trunk-mugcommander/java/source/com/mucommander/ui/main/StatusBar.java
@@ -18,6 +18,25 @@
 
 package com.mucommander.ui.main;
 
+import java.awt.Color;
+import java.awt.Dimension;
+import java.awt.Graphics;
+import java.awt.event.ComponentEvent;
+import java.awt.event.ComponentListener;
+import java.awt.event.MouseEvent;
+import java.awt.event.MouseListener;
+import java.awt.image.BufferedImage;
+import java.io.IOException;
+
+import javax.swing.Box;
+import javax.swing.BoxLayout;
+import javax.swing.Icon;
+import javax.swing.ImageIcon;
+import javax.swing.JLabel;
+import javax.swing.JPanel;
+import javax.swing.JPopupMenu;
+import javax.swing.SwingConstants;
+
 import com.mucommander.AppLogger;
 import com.mucommander.cache.FastLRUCache;
 import com.mucommander.cache.LRUCache;
@@ -39,18 +58,11 @@ import com.mucommander.ui.event.TableSelectionListener;
 import com.mucommander.ui.icon.SpinningDial;
 import com.mucommander.ui.main.table.FileTable;
 import com.mucommander.ui.main.table.FileTableModel;
-import com.mucommander.ui.theme.*;
-
-import javax.swing.*;
-import java.awt.Color;
-import java.awt.Dimension;
-import java.awt.Graphics;
-import java.awt.event.ComponentEvent;
-import java.awt.event.ComponentListener;
-import java.awt.event.MouseEvent;
-import java.awt.event.MouseListener;
-import java.awt.image.BufferedImage;
-import java.io.IOException;
+import com.mucommander.ui.theme.ColorChangedEvent;
+import com.mucommander.ui.theme.FontChangedEvent;
+import com.mucommander.ui.theme.Theme;
+import com.mucommander.ui.theme.ThemeListener;
+import com.mucommander.ui.theme.ThemeManager;
 
 
 /**
@@ -679,7 +691,7 @@ public class StatusBar extends JPanel implements Runnable, MouseListener, Active
 
                 g.setColor(c);
 
-                int freeSpaceWidth = Math.max(Math.round(freeSpacePercentage*(float)(width-2)), 1);
+				int freeSpaceWidth = Math.max(Math.round(freeSpacePercentage * (width - 2)), 1);
                 g.fillRect(1, 1, freeSpaceWidth + 1, height - 2);
 
                 // Fill background
